+++ lib/potassium/assets/active_admin/admin_user_policy.rb
@@ -1,2 +1,0 @@
-class AdminUserPolicy < ApplicationPolicy
-end