+++ lib/potassium/assets/active_admin/comment_policy.rb
@@ -1,2 +1,0 @@
-class ActiveAdmin::CommentPolicy < ApplicationPolicy
-end