+++ lib/potassium/assets/active_admin/policies/admin_user_policy.rb
@@ -1,0 +1,2 @@
+class BackOffice::AdminUserPolicy < BackOffice::DefaultPolicy
+end