+++ lib/potassium/assets/active_admin/pundit_page_policy.rb
@@ -1,5 +1,0 @@
-class ActiveAdmin::PagePolicy < ApplicationPolicy
-  def show?
-    true
-  end
-end